# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Shram API",
        "description": "API documentation for the Shram project",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/v1",
            "description": "Local API Server"
        }
    ],
    "paths": {
...
        "/owner/posts": {
            "get": {
                "tags": [
                    "Owner Posts"
                ],
                "summary": "List all posts for the authenticated owner",
                "description": "List posts endpoint",
                "operationId": "4b7a74d747dcc34a0a72fc1e4fa80e0d",
                "responses": {
                    "200": {
                        "description": "Posts retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PostListResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthorizedResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Owner Posts"
                ],
                "summary": "Create a new owner post",
                "description": "Create post endpoint",
                "operationId": "aec1a93f1fe9c9aca899e9f29a32a8e5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreatePostRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Owner post created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PostCreationResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthorizedResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/owner/posts/{id}": {
            "get": {
                "tags": [
                    "Owner Posts"
                ],
                "summary": "Show a specific post",
                "description": "Show post endpoint",
                "operationId": "b59c8ad912dfb6e60a4cf78b2eba8889",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Post ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Post retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PostShowResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthorizedResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Post not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFoundResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Owner Posts"
                ],
                "summary": "Update a specific post",
                "description": "Update post endpoint",
                "operationId": "e40149a9ab7e473cb0ba898eddc7cf28",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Post ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdatePostRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Post updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PostUpdateResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthorizedResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Post not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFoundResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Owner Posts"
                ],
                "summary": "Delete a specific post",
                "description": "Delete post endpoint",
                "operationId": "410093cb30ef32d1818df5621e3cfcc5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Post ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Post deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PostDeleteResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthorizedResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Post not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFoundResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "OwnerPost": {
                "properties": {
                    "id": {
                        "description": "Owner Post schema",
                        "type": "integer",
                        "example": 1
                    },
                    "owner_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Need 5 Labourers for Construction"
                    },
                    "description": {
                        "type": "string",
                        "example": "Detailed job description here"
                    },
                    "required_labours": {
                        "type": "integer",
                        "example": 5
                    },
                    "location": {
                        "type": "string",
                        "example": "Mumbai"
                    },
                    "start_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-08-20"
                    },
                    "end_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-08-25"
                    },
                    "work_type": {
                        "type": "string",
                        "enum": [
                            "daily",
                            "hourly"
                        ],
                        "example": "daily"
                    },
                    "wage_per_day": {
                        "type": "number",
                        "format": "float",
                        "example": 500
                    },
                    "wage_per_hour": {
                        "type": "number",
                        "format": "float",
                        "example": 70
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "open",
                            "closed",
                            "in_progress"
                        ],
                        "example": "open"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-08-13T08:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-08-13T08:00:00Z"
                    }
                },
                "type": "object"
            },
            "CreatePostRequest": {
                "required": [
                    "title",
                    "required_labours",
                    "location",
                    "work_type"
                ],
                "properties": {
                    "title": {
                        "description": "Create Post Request schema",
                        "type": "string",
                        "maxLength": 150,
                        "example": "Need 5 Labourers for Construction"
                    },
                    "description": {
                        "type": "string",
                        "example": "Detailed job description here",
                        "nullable": true
                    },
                    "required_labours": {
                        "type": "integer",
                        "minimum": 1,
                        "example": 5
                    },
                    "location": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "Mumbai"
                    },
                    "start_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-08-20",
                        "nullable": true
                    },
                    "end_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-08-25",
                        "nullable": true
                    },
                    "work_type": {
                        "type": "string",
                        "enum": [
                            "daily",
                            "hourly"
                        ],
                        "example": "daily"
                    },
                    "wage_per_day": {
                        "type": "number",
                        "format": "float",
                        "example": 500,
                        "nullable": true
                    },
                    "wage_per_hour": {
                        "type": "number",
                        "format": "float",
                        "example": 70,
                        "nullable": true
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "open",
                            "closed",
                            "in_progress"
                        ],
                        "example": "open",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "UpdatePostRequest": {
                "properties": {
                    "title": {
                        "description": "Update Post Request schema",
                        "type": "string",
                        "maxLength": 150,
                        "example": "Updated Job Title"
                    },
                    "description": {
                        "type": "string",
                        "example": "Updated job description",
                        "nullable": true
                    },
                    "required_labours": {
                        "type": "integer",
                        "minimum": 1,
                        "example": 10
                    },
                    "location": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "Delhi"
                    },
                    "start_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-09-01",
                        "nullable": true
                    },
                    "end_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-09-10",
                        "nullable": true
                    },
                    "work_type": {
                        "type": "string",
                        "enum": [
                            "daily",
                            "hourly"
                        ],
                        "example": "hourly"
                    },
                    "wage_per_day": {
                        "type": "number",
                        "format": "float",
                        "example": 600,
                        "nullable": true
                    },
                    "wage_per_hour": {
                        "type": "number",
                        "format": "float",
                        "example": 80,
                        "nullable": true
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "open",
                            "closed",
                            "in_progress"
                        ],
                        "example": "in_progress",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "ErrorResponse": {
                "properties": {
                    "status": {
                        "description": "Error response schema",
                        "type": "boolean",
                        "example": false
                    },
                    "code": {
                        "type": "string",
                        "example": "422"
                    },
                    "message": {
                        "type": "string",
                        "example": "Validation failed"
                    },
                    "errors": {
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "UnauthorizedResponse": {
                "properties": {
                    "status": {
                        "description": "Unauthorized response schema",
                        "type": "boolean",
                        "example": false
                    },
                    "error": {
                        "type": "string",
                        "example": "Unauthorized"
                    },
                    "message": {
                        "type": "string",
                        "example": "The provided credentials are incorrect"
                    }
                },
                "type": "object"
            },
            "NotFoundResponse": {
                "properties": {
                    "status": {
                        "description": "Not found response schema",
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Post not found"
                    }
                },
                "type": "object"
            },
...
            "PostCreationResponse": {
                "properties": {
                    "status": {
                        "description": "Post creation success response",
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "Post created successfully"
                    },
                    "data": {
                        "$ref": "#/components/schemas/OwnerPost"
                    }
                },
                "type": "object"
            },
            "PostListResponse": {
                "properties": {
                    "status": {
                        "description": "Post list response",
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "Posts retrieved successfully"
                    },
                    "data": {
                        "properties": {
                            "current_page": {
                                "type": "integer",
                                "example": 1
                            },
                            "data": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/OwnerPost"
                                }
                            },
                            "per_page": {
                                "type": "integer",
                                "example": 10
                            },
                            "total": {
                                "type": "integer",
                                "example": 25
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "PostShowResponse": {
                "properties": {
                    "status": {
                        "description": "Post show response",
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "Post retrieved successfully"
                    },
                    "data": {
                        "$ref": "#/components/schemas/OwnerPost"
                    }
                },
                "type": "object"
            },
            "PostUpdateResponse": {
                "properties": {
                    "status": {
                        "description": "Post update response",
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "Post updated successfully"
                    },
                    "data": {
                        "$ref": "#/components/schemas/OwnerPost"
                    }
                },
                "type": "object"
            },
            "PostDeleteResponse": {
                "properties": {
                    "status": {
                        "description": "Post delete response",
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "Post deleted successfully"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter your bearer token in the format **Bearer &lt;token&gt;**",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}